// milestory-frontend/src/api/model/getFoundationStatus200Response.ts
/**
 * Milestory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetFoundationStatus200ResponseMigration } from './getFoundationStatus200ResponseMigration';
import { GetFoundationStatus200ResponseDatabase } from './getFoundationStatus200ResponseDatabase';


export interface GetFoundationStatus200Response { 
    headline: string;
    summary: string;
    mode: GetFoundationStatus200Response.ModeEnum;
    apiVersion: string;
    database: GetFoundationStatus200ResponseDatabase;
    migration: GetFoundationStatus200ResponseMigration;
    generatedAt: string;
    notes: Array<string>;
}
export namespace GetFoundationStatus200Response {
    export const ModeEnum = {
        Ready: 'ready',
        Empty: 'empty',
        Degraded: 'degraded'
    } as const;
    export type ModeEnum = typeof ModeEnum[keyof typeof ModeEnum];
}